ci: php 8.5 (#7585)

+++ src/Laravel/Tests/GraphQlTest.php
@@ -83,7 +83,7 @@ class GraphQlTest extends TestCase
         // Create books in reverse alphabetical order to test the 'asc' order
         BookFactory::new()
             ->count(10)
-            ->sequence(static fn ($sequence) => ['name' => \chr(122 - $sequence->index)]) // ASCII codes starting from 'z'
+            ->sequence(static fn ($sequence) => ['name' => \chr((int) (122 - $sequence->index))]) // ASCII codes starting from 'z'
             ->has(AuthorFactory::new())
             ->create();
